refactor:extract joyride from userinfo index.tsx
